Refactor code to add more OOP
