+++ src/main/java/processor/Processor.java
@@ -5,21 +5,20 @@ import exception.DukeException;
 import models.Command;
 import models.Deadline;
 import models.Event;
-import models.TaskList;
 import models.Todo;
-import util.Output;
+import storage.Storage;
+import ui.Ui;
 
 import java.util.List;
 
 
 import java.time.format.DateTimeParseException;
 import java.time.LocalDate;
-import java.util.List;
 
 
 public class Processor implements IProcessor {
 
-    private TaskList list = new TaskList(Setting.FILE_DIRECTORY, Setting.FILE_NAME);
+    private Storage storage = new Storage(Setting.FILE_DIRECTORY, Setting.FILE_NAME);
 
     @Override
     public void processCommand(Command command, List<String> arguments) {
@@ -50,7 +49,7 @@ public class Processor implements IProcessor {
                 if (arguments.size() == 0) {
                     throw new DukeException("Todo description cannot be empty");
                 }
-                this.list.addTask(new Todo(String.join(" ", arguments)));
+                this.storage.addTask(new Todo(String.join(" ", arguments)));
             } else if (type.equals("deadline")) {
                 arguments.remove(0);
                 if (arguments.size() == 0) {
@@ -62,7 +61,7 @@ public class Processor implements IProcessor {
                     throw new DukeException("Deadline command must have /by specified");
                 }
                 LocalDate time = LocalDate.parse(input[1].trim());
-                this.list.addTask(new Deadline(input[0], time));
+                this.storage.addTask(new Deadline(input[0], time));
             } else if (type.equals("event")) {
                 arguments.remove(0);
                 if (arguments.size() == 0) {
@@ -74,31 +73,31 @@ public class Processor implements IProcessor {
                     throw new DukeException("Event command must have /at specified");
                 }
                 LocalDate time = LocalDate.parse(input[1].trim());
-                this.list.addTask(new Event(input[0], time));
+                this.storage.addTask(new Event(input[0], time));
             } else {
                 throw new DukeException("I don't understand:(");
             }
-            Output.print("Got it. I've added this task:\n   " + this.list.getLastTask() + "\nNow you have " + this.list.getSize() + " tasks in the list.");
+            Ui.print("Got it. I've added this task:\n   " + this.storage.getLastTask() + "\nNow you have " + this.storage.getSize() + " tasks in the list.");
         } catch (DukeException e) {
-            Output.print(e.getMessage());
+            Ui.print(e.getMessage());
         } catch (DateTimeParseException e) {
-            Output.print(e.getMessage());
+            Ui.print(e.getMessage());
         }
     }
 
     @Override
     public void processList() {
-        Output.print(this.list.toString());
+        Ui.print(this.storage.toString());
     }
 
     @Override
     public void processDone(String index) {
         try {
             int i = Integer.parseInt(index);
-            this.list.setDone(i - 1);
-            Output.print("Nice! I've marked this task as done:\n   " + this.list.getTask(i - 1));
+            this.storage.setDone(i - 1);
+            Ui.print("Nice! I've marked this task as done:\n   " + this.storage.getTask(i - 1));
         } catch(DukeException e) {
-            Output.print(e.getMessage());
+            Ui.print(e.getMessage());
         }
     }
 
@@ -106,15 +105,15 @@ public class Processor implements IProcessor {
     public void processDelete(String index) {
         try {
             int i = Integer.parseInt(index);
-            String result = this.list.deleteTask(i - 1);
-            Output.print("Got it! I've removed this task:\n   " + result + "\nNow you have " + this.list.getSize() + " tasks in the list.");
+            String result = this.storage.deleteTask(i - 1);
+            Ui.print("Got it! I've removed this task:\n   " + result + "\nNow you have " + this.storage.getSize() + " tasks in the list.");
         } catch (DukeException e) {
-            Output.print(e.getMessage());
+            Ui.print(e.getMessage());
         }
     }
 
     @Override
     public void processBye() {
-        Output.print("Bye. Please meet me again later!");
+        Ui.print("Bye. Please meet me again later!");
     }
 }
