+++ src/main/java/ui/Ui.java
@@ -1,6 +1,6 @@
-package util;
+package ui;
 
-public class Output {
+public class Ui {
     public static void print(String line) {
         System.out.println("\t" + line.replace("\n", "\n\t"));
         System.out.println("\t" + "____________________________________________________________\n");
